Pad more 128 bytes when necessary to complete a 256 bytes sector
As the firmware image is transferred in blocks of 128 bytes, and the
sector size is 256 bytes, sometimes the firmware image may be divided
in a odd number of 128 bytes blocks, preventing the last flash sector
of been written. This causes firmware corruption.

I've added a check to ensure that a even number of 128 bytes blocks
are always sent.